Beach-Scene-October-24-2018
As of October 24, this is the current design for the beach scene. We got the basic design of the island in. We're working on adding sound, adding a first person perspective, and plant life on it. To access the actual scene to go Assests > Beach Scene.unity > Once opened the scene on Unity, click Project > Scenes > Free Island Collection > Sand > Scene 5. It should be an elevated island surronded by water.